refactor reviews to ratingsAndReviews
